requirements.txt
requests
openai
sounddevice
openai-whisper
wavio
gtts
playsound
TTS
PyObjC

# Test and lint
pytest
pylint
flake8
